Updated Merge Sorted Array with time and space complexity analysis
