+++ readMe.txt
@@ -1,0 +1,10 @@
+Hi Everyone
+
+This repo consists of the solution for Leetcode 150 top interview question series
+
+I would be giving a brief idea behind the approach I used and often I will add some fun alternates approaches as well.
+The time and space complexity analysis will also be provided.
+
+Looking forward for your feedback and always happy to help.
+
+
